+++ src/main/java/com/mrh0/createaddition/energy/network/EnergyNetwork.java
@@ -11,12 +11,12 @@ import net.minecraft.world.level.Level;
 public class EnergyNetwork {
 	private int id;
 	// Input
-	private int inBuff;
-	private int inDemand;
+	private long inBuff;
+	private long inDemand;
 	// Output
-	private int outBuff;
-	private int outBuffRetained;
-	private int outDemand;
+	private long outBuff;
+	private long outBuffRetained;
+	private long outDemand;
 	private boolean valid;
 	
 	private int pulled = 0;
@@ -35,29 +35,29 @@ public class EnergyNetwork {
 		EnergyNetworkManager.instances.get(world).add(this);
 	}
 
-	public int getMaxBuff() {
+	public long getMaxBuff() {
 		return Math.min(nodeCount * (outDemand + inDemand * 2 + 10), CommonConfig.CONNECTOR_NETWORK_INTERNAL_BUFFER.get());
 	}
-	
+
 	public void tick(int index) {
 		this.id = index;
-		int t = outBuff;
+		long t = outBuff;
 		outBuff = inBuff;
 		outBuffRetained = outBuff;
 		inBuff = t;
 		outDemand = inDemand;
 		inDemand = 0;
-				
+
 		pulled = 0;
 		pushed = 0;
 	}
-	
-	public int getBuff() {
+
+	public long getBuff() {
 		return outBuffRetained;
 	}
 
 	// Returns the amount of energy pushed to network
-	public int push(int energy, boolean simulate) {
+	public long push(long energy, boolean simulate) {
 		energy = Math.min(getMaxBuff() - inBuff, energy);
 		energy = Math.max(energy, 0);
 		if (!simulate) {
@@ -67,30 +67,30 @@ public class EnergyNetwork {
 		return energy;
 	}
 
-	public int push(int energy) {
+	public long push(long energy) {
 		return push(energy, false);
 	}
-	
-	public int demand(int demand) {
+
+	public long demand(long demand) {
 		this.inDemand += demand;
 		return demand;
 	}
-	
-	public int getDemand() {
+
+	public long getDemand() {
 		return outDemand;
 	}
-	
+
 	public int getPulled() {
 		return pulled;
 	}
-	
+
 	public int getPushed() {
 		return pushed;
 	}
 
 	// Returns amount of energy pulled from network
-	public int pull(int energy, boolean simulate) {
-		int r = Math.max(Math.min(energy, outBuff), 0);
+	public long pull(long energy, boolean simulate) {
+		long r = Math.max(Math.min(energy, outBuff), 0);
 		if (!simulate) {
 			outBuff -= r;
 			pulled += r;
@@ -98,16 +98,16 @@ public class EnergyNetwork {
 		return r;
 	}
 
-	public int pull(int max) {
+	public long pull(long max) {
 		return pull(max, false);
 	}
-	
+
 	public static EnergyNetwork nextNode(Level level, EnergyNetwork en, Map<String, IWireNode> visited, IWireNode current, int index) {
 		if (visited.containsKey(posKey(current.getPos(), index))) return null; // should never matter?
 		current.setNetwork(index, en);
 		visited.put(posKey(current.getPos(), index), current);
 		en.nodeCount++;
-		
+
 		for (int i = 0; i < current.getNodeCount(); i++) {
 			IWireNode next = current.getWireNode(i);
 			if (next == null) continue;
@@ -116,21 +116,21 @@ public class EnergyNetwork {
 		}
 		return en;
 	}
-	
+
 	private static String posKey(BlockPos pos, int index) {
 		return pos.getX()+","+pos.getY()+","+pos.getZ()+":"+index;
 	}
-	
+
 	public void invalidate() {
 		this.valid = false;
 	}
-	
+
 	public boolean isValid() {
 		return this.valid;
 	}
-	
+
 	public void removed() {}
-	
+
 	public int getId() {
 		return id;
 	}
